Move tests into same file
